Minor bugs fixed
Fixed a bug that caused the location fetching system to fail on initial loading and improved the speed of the WHOAMI command.
