+++ public/src/ex_file/updatelog.html
@@ -1,4 +1,9 @@
 <div style="height:400px;overflow-y:scroll;padding:0 15px;">
+  <h2>August 12, 2021 (V. 01-7-0.21)</h2>
+  <ul>
+    <li>Optimized the WHOAMI profile command, it should run smoother now.</li>
+    <li>Fixed a bug that caused the location fetching system to fail on initial loading.</li>
+  </ul>
   <h2>August 11, 2021 (V. 01-6-0.21)</h2>
   <ul>
     <li>Added a settings command. You can now mask your location by adjusting the settings.</li>
