+++ public/src/js/script.js
@@ -22,7 +22,7 @@ let generator = new Generator(180, {
   minimumRatio: 3.1
 });
 //update version
-$("#version").text("V. 01-6-0.21")
+$("#version").text("V. 01-7-0.21")
 //script variables
 var access
 var vcLoaded = false;
@@ -595,101 +595,10 @@ function reply(val) {
         cmdHide()
         $d.append(`<blockquote id="waitingToAdd">Accessing your profile...</blockquote>`)
         addDot()
-        loadCroppie(function() {
-          var storageRef = firebase.storage()
-          storageRef.ref(`user/${firebase.auth().getUid()}/pfp/pfp.jpg`).getDownloadURL().then(onResolve, onReject);
-
-          function onResolve(foundURL) {
-            $d.append(`<br><br><div class="profile editPhoto" data-uid="${firebase.auth().getUid()}">
-                      <h3 style="text-align: center;"><strong>${displayName}'S PROFILE</strong></h3>
-                      <img class="profilePic" id="${profileListing}-profile" style="float: right; position: relative; width: 200px; height: 228px;" src="${foundURL}" />
-                      <div class="profileData" style="float: left;max-width: 250px;"><small style="text-align: left;">USERNAME:</small>
-                      <p class="highlight" style="text-align: left;">${displayName}<span style="color: rgba(159, 159, 159, 0.64);">#${tagNo}</span></p>
-                      <hr /><small style="text-align: left;">SECURITY CLEARANCE LEVEL:</small>
-                      <p class="highlight" style="text-align: left;">${clearance}</p>
-                      <hr /><small>PERSONNEL CLASSIFICATION:</small>
-                      <p class="highlight">${classification}</p>
-                      <hr /><small style="text-align: left;">STAFF TITLE:</small>
-                      <p class="highlight" style="text-align: left;">${title}</p>
-                      <hr /><small style="text-align: left;">WORKING SITE:</small>
-                      <p class="highlight" style="text-align: left;">${site}</p>
-                      <hr />
-                      <p style="text-align: left;">&nbsp;</p>
-                      </div>
-                      <div style="clear: both;">&nbsp;</div>
-                      <h3 style="text-align: center;"><strong>CONFIDENTIAL INFORMATION</strong></h3>
-                      <small style="text-align: left;">EMAIL ADDRESS:</small>
-                      <p class="highlight" style="text-align: left;">${firebase.auth().currentUser.email}</p>
-                      <hr />
-                      <small style="text-align: left;">USER ID:</small>
-                      <p class="highlight" style="text-align: left;">${tag}</p>
-                      <hr />
-                      <small style="text-align: left;">DATE OF JOINING THE FOUNDATION:</small>
-                      <p class="highlight" style="text-align: left;">${new Date(firebase.auth().currentUser.metadata.creationTime).toLocaleDateString("en-US")}</p>
-                      <hr />
-                      <small style="text-align: left;">LAST LOGIN TIME:</small>
-                      <p class="highlight" style="text-align: left;">${new Date(firebase.auth().currentUser.metadata.lastSignInTime).toLocaleString('en-US')}</p>
-                      <hr />
-                      </div><br>`)
-            $d.find(`#${profileListing}-profile`).on('load', function() {
-              if ($(this).parent(".editPhoto").is(":visible") == false) {
-                $(this).parent(".editPhoto").show()
-                $d.append(`${holder} `)
-                cmdShow()
-                btnShow()
-                scroll()
-              }
-            });
-          }
-
-          function onReject(error) {
-            if (error.code == "storage/object-not-found") {
-              $d.append(`<br><br><div class="profile editPhoto" id="${profileListing}-profile" data-uid="${firebase.auth().getUid()}">
-                       <h3 style="text-align: center;"><span><strong>${displayName}'S PROFILE</strong></span></h3>
-                       <div class="upload" style="float: right;position: relative;">
-                            Click to add a profile picture<br><b>+</b><br><small>File should not exceed 20 MB</small>
-                            <input class="file_upload" type="file" accept="image/*" onchange="parent.uploadfile(this)" onclick="this.value=null;" />
-                        </div>
-                       <div class="profileData" style="float: left;max-width: 250px;">
-                          <small style="text-align: left;">USERNAME:</small>
-                          <p class="highlight" style="text-align: left;">${displayName}<span style="color: rgba(159, 159, 159, 0.64);">#${tagNo}</span></p>
-                          <hr />
-                          <small style="text-align: left;">SECURITY CLEARANCE LEVEL:</small>
-                          <p class="highlight" style="text-align: left;">${clearance}</p>
-                          <hr />
-                          <small>PERSONNEL CLASSIFICATION:</small>
-                          <p class="highlight">${classification}</p>
-                          <hr />
-                          <small style="text-align: left;">STAFF TITLE:</small>
-                          <p class="highlight" style="text-align: left;">${title}</p>
-                          <hr />
-                          <small style="text-align: left;">WORKING SITE:</small>
-                          <p class="highlight" style="text-align: left;">${site}</p>
-                          <hr />
-                          <p style="text-align: left;">&nbsp;</p>
-                       </div>
-                       <div style="clear: both;">&nbsp;</div>
-                       <h3 style="text-align: center;"><strong>CONFIDENTIAL INFORMATION</strong></h3>
-                       <small style="text-align: left;">EMAIL ADDRESS:</small>
-                       <p class="highlight" style="text-align: left;">${firebase.auth().currentUser.email}</p>
-                       <hr />
-                       <small style="text-align: left;">USER ID:</small>
-                       <p class="highlight" style="text-align: left;">${tag}</p>
-                       <hr />
-                       <small style="text-align: left;">DATE OF JOINING THE FOUNDATION:</small>
-                       <p class="highlight" style="text-align: left;">${new Date(firebase.auth().currentUser.metadata.creationTime).toLocaleDateString("en-US")}</p>
-                       <hr />
-                       <small style="text-align: left;">LAST LOGIN TIME:</small>
-                       <p class="highlight" style="text-align: left;">${new Date(firebase.auth().currentUser.metadata.lastSignInTime).toLocaleString('en-US')}</p>
-                       <hr />
-                       </div><br>`)
-            }
-            $d.find(".profile").show()
-            $d.append(`${holder} `)
-            cmdShow()
-            btnShow()
-            scroll()
-          }
+        import( /*webpackChunkName:'locate'*/ './profile.js').then((module) => {
+          loadCroppie(function() {
+            module.whoami()
+          })
         })
       } else {
         appendNoLogin()
@@ -834,12 +743,20 @@ window.locationMasking = (con) => {
     $("#ip, #location, #tel").html('<span style="color:#EA3546"><span style="font-weight:bold">ⓘ</span> [REQUEST REFUSED]</span>')
   } else {
     settings["checkLocation"] = false
-    if (place != "") {
-      $("#location").text(place)
-      $("#ip").html(`${ip}`)
-      $("#tel").text(`${tele}`)
-      locationGet = true
-      window.displayLoc = place.toUpperCase()
+    checkPlaceLoaded()
+
+    function checkPlaceLoaded() {
+      if (typeof place !== 'undefined') {
+        if (place != "") {
+          $("#location").text(place)
+          $("#ip").html(`${ip}`)
+          $("#tel").text(`${tele}`)
+          locationGet = true
+          window.displayLoc = place.toUpperCase()
+        }
+        return;
+      }
+      window.setTimeout(checkPlaceLoaded, 100)
     }
   }
 }
@@ -975,7 +892,7 @@ function loadCroppie(callback) {
         isProfileLoaded = true
       }
     });
-    $.getScript("/__/firebase/8.7.1/firebase-app-check.js", function() {
+    $.getScript("/__/firebase/8.8.0/firebase-app-check.js", function() {
       //appCheck setup
       const appCheck = firebase.appCheck();
       appCheck.activate('6Lch95QbAAAAAKydxDgt3zyqBGtAt9WWQ2-qafVi');
